DEMRUM-4710: Formatting issues
